Use core gif encoder.
